backend/requirements.txt
fastapi
uvicorn
pinecone
sentence-transformers
python-dotenv